auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.0
python-dotenv
python-jose[cryptography]
httpx